forwarding: extract source-NAT + interface-NAT helpers into nat.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the source-NAT
flow matchers (nat_scope_ctx_for_flow, match_source_nat_for_flow*,
match_source_nat_for_flow_result_at) and the interface-NAT local
resolution helpers (interface_nat_local_resolution*,
should_block_tunnel_interface_nat_session_miss) into forwarding/nat.rs.

Functions are moved verbatim; no logic, signature, or hot-path change.
pub(super) becomes pub(in crate::afxdp) so the mod.rs glob re-export
keeps all call sites resolving. Preserved the #[cfg_attr(not(test),
allow(dead_code))] / #[allow(clippy::too_many_arguments)] attributes.

Validation: cargo build clean (0 errors).
